indexer: tx_affected_objects ingestion-based backfill
## Description

Re-implement the `tx_affected_objects` backfill to use the
ingestion-based backfill system, as the previous approach was too slow.

## Test plan

- Run the indexer locally, against a remote store.
- Create a copy of its `tx_affected_objects` table and create a fresh
  table to backfil into.
- Find the backfill high watermark:

```sql
SELECT MAX(sequence_number) FROM checkpoints;
```

- Run the backfill against the same remote store, between `0` and the
  high watermark.
- Confirm that the table you created with backfill has the same contents
  as the backup:

```sql
-- All of these should be the same count
SELECT COUNT(1) FROM tx_affected_objects;

SELECT COUNT(1) FROM tx_affected_objects_backup;

SELECT
    COUNT(1)
FROM
    tx_affected_objects
INNER JOIN
    tx_affected_objects_backup
USING (
    affected,
    tx_sequence_number,
    sender
);
```
