+++ crates/sui-indexer/src/backfill/backfill_instances/mod.rs
@@ -3,6 +3,7 @@
 
 use crate::backfill::backfill_instances::ingestion_backfills::ingestion_backfill_task::IngestionBackfillTask;
 use crate::backfill::backfill_instances::ingestion_backfills::raw_checkpoints::RawCheckpointsBackFill;
+use crate::backfill::backfill_instances::ingestion_backfills::tx_affected_objects::TxAffectedObjectsBackfill;
 use crate::backfill::backfill_task::BackfillTask;
 use crate::backfill::{BackfillTaskKind, IngestionBackfillKind};
 use std::sync::Arc;
@@ -11,7 +12,6 @@ use sui_types::messages_checkpoint::CheckpointSequenceNumber;
 mod ingestion_backfills;
 mod sql_backfill;
 mod system_state_summary_json;
-mod tx_affected_objects;
 
 pub async fn get_backfill_task(
     kind: BackfillTaskKind,
@@ -20,9 +20,6 @@ pub async fn get_backfill_task(
     match kind {
         BackfillTaskKind::SystemStateSummaryJson => {
             Arc::new(system_state_summary_json::SystemStateSummaryJsonBackfill)
-        }
-        BackfillTaskKind::TxAffectedObjects => {
-            Arc::new(tx_affected_objects::TxAffectedObjectsBackfill)
         }
         BackfillTaskKind::Sql { sql, key_column } => {
             Arc::new(sql_backfill::SqlBackFill::new(sql, key_column))
@@ -33,6 +30,13 @@ pub async fn get_backfill_task(
         } => match kind {
             IngestionBackfillKind::RawCheckpoints => Arc::new(
                 IngestionBackfillTask::<RawCheckpointsBackFill>::new(
+                    remote_store_url,
+                    range_start as CheckpointSequenceNumber,
+                )
+                .await,
+            ),
+            IngestionBackfillKind::TxAffectedObjects => Arc::new(
+                IngestionBackfillTask::<TxAffectedObjectsBackfill>::new(
                     remote_store_url,
                     range_start as CheckpointSequenceNumber,
                 )
