+++ crates/sui-indexer/src/backfill/mod.rs
@@ -10,7 +10,6 @@ pub mod backfill_task;
 #[derive(Subcommand, Clone, Debug)]
 pub enum BackfillTaskKind {
     SystemStateSummaryJson,
-    TxAffectedObjects,
     /// \sql is the SQL string to run, appended with the range between the start and end,
     /// as well as conflict resolution (see sql_backfill.rs).
     /// \key_column is the primary key column to use for the range.
@@ -31,4 +30,5 @@ pub enum BackfillTaskKind {
 #[derive(ValueEnum, Clone, Debug)]
 pub enum IngestionBackfillKind {
     RawCheckpoints,
+    TxAffectedObjects,
 }
